Fix compilation warnings
Former-commit-id: 432e240cae044dc044556e237cebfd86117edc5a
Former-commit-id: b87b349c572d513bfecce075850843b8b0e78f29 [formerly 56dd4c439c64c8956a6a244e8297fb935810019c]
Former-commit-id: 795112c70d9dc66858b3a17cdf66138daf454a06
Former-commit-id: 15a402f3941b7cdfeef83b7254bd0b8bd3293235 [formerly 3d97a093ba570c9aaa8a7ba11628a58c73c5d255]
Former-commit-id: a661b9f23eefc9e2bb5855378b19b28861c72066
Former-commit-id: 84a5316193e720589a8d1239f55f203ad0c1a09b
Former-commit-id: 242442de952ea47f2d4b2bfe50dcc3b63343f354

+++ src/C_lassmooth.cpp
@@ -45,7 +45,7 @@ NumericVector C_lassmooth(S4 las, double size, int method = 1, int shape = 1, do
   NumericVector Y = data["Y"];
   NumericVector Z = data["Z"];
 
-  long n = X.length();
+  unsigned int n = X.length();
   double half_res = size / 2;
   double twosquaresigma = 2*sigma*sigma;
   double twosquaresigmapi = twosquaresigma * PI;
@@ -57,7 +57,7 @@ NumericVector C_lassmooth(S4 las, double size, int method = 1, int shape = 1, do
 
   Progress p(n, false);
 
-  for (long i = 0 ; i < n ; i++)
+  for (unsigned int i = 0 ; i < n ; i++)
   {
     std::vector<Point*> pts;
 
@@ -70,7 +70,7 @@ NumericVector C_lassmooth(S4 las, double size, int method = 1, int shape = 1, do
     double ztot = 0;
     double wtot = 0;
 
-    for(long j = 0 ; j < pts.size() ; j++)
+    for(unsigned int j = 0 ; j < pts.size() ; j++)
     {
       if (method == 1)
       {
